refactor(tests/system/{common,nar-info-querying}): extract subprocess utils
