+++ tests/system/common/src/lib.rs
@@ -1,1 +1,1 @@
-
+pub mod subprocess;
