+++ tests/system/nar-info-querying/Cargo.toml
@@ -12,3 +12,5 @@ reqwest = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true }
 which = { workspace = true }
+
+selector4nix-system-test-common = { path = "../common" }
